rename node and consolidate tests
