+++ src/vellum/workflows/nodes/displayable/tool_calling_node/tests/test_utils.py
@@ -1,7 +1,9 @@
+import pytest
+
 from vellum.workflows import BaseWorkflow
 from vellum.workflows.inputs.base import BaseInputs
 from vellum.workflows.nodes.bases import BaseNode
-from vellum.workflows.nodes.displayable.tool_calling_node.utils import create_function_node, get_function_name
+from vellum.workflows.nodes.displayable.tool_calling_node.utils import get_function_name
 from vellum.workflows.outputs.base import BaseOutputs
 from vellum.workflows.state.base import BaseState
 from vellum.workflows.types.definition import ComposioToolDefinition, DeploymentDefinition
@@ -58,80 +60,19 @@ def test_get_function_name_subworkflow_deployment_uuid():
     assert result == "57f09bebb46340e0bf9ec972e664352f"
 
 
-def test_get_function_name_composio_tool_definition():
-    """Test ComposioToolDefinition function name generation."""
-    composio_tool = ComposioToolDefinition(
-        toolkit="GITHUB", action="GITHUB_CREATE_AN_ISSUE", description="Create a new issue in a GitHub repository"
-    )
+@pytest.mark.parametrize(
+    "toolkit,action,description,expected_result",
+    [
+        ("SLACK", "SLACK_SEND_MESSAGE", "Send message to Slack", "slack_send_message"),
+        ("GMAIL", "GMAIL_CREATE_EMAIL_DRAFT", "Create Gmail draft", "gmail_create_email_draft"),
+    ],
+)
+def test_get_function_name_composio_tool_definition_various_toolkits(
+    toolkit: str, action: str, description: str, expected_result: str
+):
+    """Test ComposioToolDefinition function name generation with various toolkits."""
+    composio_tool = ComposioToolDefinition(toolkit=toolkit, action=action, description=description)
 
     result = get_function_name(composio_tool)
 
-    assert result == "github_github_create_an_issue"
-
-
-def test_get_function_name_composio_tool_definition_various_toolkits():
-    """Test ComposioToolDefinition function name generation with various toolkits."""
-    slack_tool = ComposioToolDefinition(
-        toolkit="SLACK", action="SLACK_SEND_MESSAGE", description="Send message to Slack"
-    )
-
-    gmail_tool = ComposioToolDefinition(
-        toolkit="GMAIL", action="GMAIL_CREATE_EMAIL_DRAFT", description="Create Gmail draft"
-    )
-
-    assert get_function_name(slack_tool) == "slack_slack_send_message"
-    assert get_function_name(gmail_tool) == "gmail_gmail_create_email_draft"
-
-
-def test_create_function_node_with_composio_tool_definition():
-    """Test that create_function_node creates ComposioFunctionNode for ComposioToolDefinition."""
-    from unittest.mock import Mock
-
-    composio_tool = ComposioToolDefinition(
-        toolkit="GITHUB", action="GITHUB_CREATE_AN_ISSUE", description="Create a new issue in a GitHub repository"
-    )
-
-    # Mock tool router node
-    mock_tool_router_node = Mock()
-    mock_tool_router_node.Outputs.results = "mock_results"
-
-    # Create function node
-    function_node_class = create_function_node(composio_tool, mock_tool_router_node)
-
-    # Check that it returns a class
-    assert isinstance(function_node_class, type)
-
-    # Check that the class name is correct
-    assert function_node_class.__name__ == "ComposioFunctionNode_github_github_create_an_issue"
-
-    # Check that it has the correct attributes
-    assert hasattr(function_node_class, "composio_tool")
-    assert hasattr(function_node_class, "function_call_output")
-
-    # Check that the attributes exist (they are wrapped in NodeReference objects)
-    assert hasattr(function_node_class.composio_tool, "__class__")
-    assert hasattr(function_node_class.function_call_output, "__class__")
-
-
-def test_create_function_node_composio_tool_inheritance():
-    """Test that ComposioFunctionNode created by create_function_node has correct inheritance."""
-    from unittest.mock import Mock
-
-    from vellum.workflows.nodes.displayable.tool_calling_node.utils import ComposioFunctionNode
-
-    composio_tool = ComposioToolDefinition(
-        toolkit="SLACK", action="SLACK_SEND_MESSAGE", description="Send message to Slack"
-    )
-
-    # Mock tool router node
-    mock_tool_router_node = Mock()
-    mock_tool_router_node.Outputs.results = "mock_results"
-
-    # Create function node
-    function_node_class = create_function_node(composio_tool, mock_tool_router_node)
-
-    # Check that it's a subclass of ComposioFunctionNode
-    assert issubclass(function_node_class, ComposioFunctionNode)
-
-    # Check that it's a subclass of BaseNode (through ComposioFunctionNode)
-    assert issubclass(function_node_class, BaseNode)
+    assert result == expected_result
